Implementação do Individuo tipo function com representação binária
Por Pedro Isidoro e Aurélien Mota, com ajuda de Paulo Chorinca.

+++ v0.4/GeneticAlgoritms/src/utils/PopulationUtils.java
@@ -230,7 +230,7 @@ public class PopulationUtils {
         return __newPopulation;
     }
 
-    public static Collection<Individual> getUniqueIndividuals(Population population, int fitness) {
+    public static Collection<Individual> getUniqueIndividuals(Population population, double fitness) {
         //Ordenação
         TreeSet<Individual> __population = new TreeSet<Individual>(new ComparatorIndividual());
